提交文件

+++ cloud-provider-payment8001/src/main/java/com/atguigu/cloud/dao/PaymentDao.java
@@ -3,11 +3,13 @@ package com.atguigu.cloud.dao;
 import com.atguigu.springcloud.entities.Payment;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
+import org.apache.ibatis.annotations.Select;
 
 @Mapper
 public interface PaymentDao {
+
      int create(Payment payment);
 
-     Payment getPaymentById(@Param("id") Long id);
+     Payment getPaymentById(@Param("id") int id);
 
 }
