+++ cloud-provider-payment8002/src/main/java/com/atguigu/cloud/dao/PaymentDao.java
@@ -8,6 +8,6 @@ import org.apache.ibatis.annotations.Param;
 public interface PaymentDao {
      int create(Payment payment);
 
-     Payment getPaymentById(@Param("id") Long id);
+     Payment getPaymentById(@Param("id") int id);
 
 }
